Compile on double build and test harness movement
